+++ CHANGELOG.md
@@ -21,7 +21,7 @@ All notable changes to this project will be documented in this file.
 - Replace single thread executor with synchronized keyword. (#633)
 - Move contribution status checks from `iexec-commons-poco`. (#636)
 - Use `BlockchainAdapterService` from `iexec-blockchain-adapter-api-library`. (#641)
-- `ResultRepositoryConfiguration` and `WorkerConfiguration` classes are now immutable with `@Data` lombok annotation. (#650)
+- `ResultRepositoryConfiguration` and `WorkerConfiguration` classes are now immutable with `@Value` lombok annotation. (#650)
 
 ### Bug Fixes